+++ nds-client/src/main/java/br/com/abril/nds/service/impl/FecharDiaServiceImpl.java
@@ -363,7 +363,7 @@ public class FecharDiaServiceImpl implements FecharDiaService {
      */
     @Override
     @Transactional(readOnly = true)
-    public int contarDividasReceberEm(Date data) {
+    public long contarDividasReceberEm(Date data) {
         return dividaService.contarDividasReceberEm(data);
     }
 
@@ -372,7 +372,7 @@ public class FecharDiaServiceImpl implements FecharDiaService {
      */
     @Override
     @Transactional(readOnly = true)
-    public int contarDividasVencerApos(Date data) {
+    public long contarDividasVencerApos(Date data) {
         return dividaService.contarDividasVencerApos(data);
     }
